Improving the comments for the app icon setting.
There is a lot of confusion over what is the right way to set app icons. I myself struggled with this setting because the default icons generated by phonegap are also named "icon.png". When I placed an `icon.png` inside `img` folder, it didn't work. It's important to make it explicit in the documentation that the setting is taken relative the `config.xml` location.