remove bithound
http://bithound.io/

```
We would like to personally extend a big thank you to all our customers who have supported us since we first came online in 2014. As with all startups we have had our ups and downs, but we truly valued the journey.

We followed our passion trying to bring a service to the market that would help individuals and companies make better decisions when it came to the building and delivery of their software. Our goal was to help entire development teams get on the same page by gaining visibility into code quality and security.

We are saddened to say that we were not successful in proving the viability of this business model over the 4 year life-span of bitHound. As such we were forced to make the very tough decision to windup the company and discontinue our analysis service.

We would like to thank you once again and wish you success in all of your current and future projects.
The bitHound Team
```